votes added
still a bit of a work in progress as placing a vote does not update the
current view of a vote as there is no state. Also bits of the
implementation are coming through with things like questionsQuiz being
seen in the front end. A couple of opitions here:

1) add state to votes which can be set by Quiz query and Vote mutation
2) switch to using Apollo cache - sounds like the best option

Also:
1) change to have votes on questions with no mention of questionsQuiz(zes)
on the front end

Next:
1) look at the after hook to finalise a vote and set it on
QuestionsQuizzes table? in which case it would need to be visible to
frontend as a Vote table?
2) polling or synchronising only on the state of Questions Quiz, ie
Votes?
3) a state of when all votes are in and the Quiz can start, another
after hook?